requirements.txt
oauthlib
requests
requests_oauthlib
python-librsync
